gkr 33109249 26/4/2023
[new file] Enemy.java:
-New abstract class for enemy. This is inherited by all the other enemy characters

Most of the changes from LoneWolf is imported into Enemy abstract class with some more fixes.
Within playTurn() in Enemy class, checks all the exits and sees if there is another actor there. If there is an actor, checks if it's the same type, is so, puts attackBehaviour on that actor.

If actor has HOSTILE_TO_ENEMY, it follows the actor forever.

[new files that inherit enemy] LoneWolfEnemy, GiantCrab, HeavySkeletalSwordsman

[packages] actions, actors, behaviours

NOTES:
Most of enemies complete. Now need to add utilisation of weapons by both enemies and player.

I think there needs to be some weaponattack action?

[things to be done (off the top of my head)]
-rune drop after actor death. (HAVE TO USE DEATHACTION)
-weapons + weapon attack action?
-change UML diagram + design rationale
-Potentially delete LoneWolf class
-Lost site of grace, and game reset
-Consumable items (consume action?)
-Way more I can't really think of