Addition of ConsumableItem, ConsumeAction, ConsumeHeal, FlaskOfCrimsonTears. Changes to AreaAttack and AttackAction, ReviveAction
ConsumableItem: New abstract class that represents and item that can be consumed with limited usages.

ConsumeAction: A new child class of Action which is used to consume ConsumableItem.

ConsumeHeal: An interface which describes the ability for a ConsumableItem to heal.

FlaskOfCrimsonTears: A ConsumableItem that implements ConsumeHeal. Works.

AreaAttackAction: Slightly edit return message after death for Pileofbones.